requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.25.0
Django==3.2.3
django-bootstrap4==3.0.1
django-phonenumber-field==5.1.0
django-quill==0.1.0
django-quill-editor==0.1.22
django-tinymce==3.3.0
importlib-metadata==2.1.1
pkg-resources==0.0.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
